requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
scikit-learn
prophet
statsmodels
dowhy
DoubleML
pulp
shap
pillow
PyPDF2
opencv-python
pytesseract
torch
torchvision
pdfplumber
spacy
pytesseract
sumy
python-magic
